small change to make the mwag not delete from

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/service/GroupsServerService.java
@@ -161,7 +161,9 @@ public class GroupsServerService extends GroupsServiceImplBase {
 
             // If this user from the groupToRemoveFrom is the actual user requested for removal.
             if ((request.getUserIdsList()).contains(user.getId())) {
-                groupMembershipRepo.deleteByRegisteredGroupsAndRegisteredGroupUser(groupToRemoveFrom, user);
+                if (!groupToRemoveFrom.getGroupShortName().equals(MWAG_GROUP_NAME_SHORT)) {
+                    groupMembershipRepo.deleteByRegisteredGroupsAndRegisteredGroupUser(groupToRemoveFrom, user);
+                }
 
                 //if the user has no groups left, add the to the MWAG
                 if (user.getGroups().isEmpty()) {
